+++ Domain/Entities/Phone.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Domain.Models
+namespace Domain.Entities
 {
     public class Phone
     {
